backend/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pymongo==4.8.0
requests==2.32.3
langgraph==0.2.35
openai>=1.40.0
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
rapidfuzz==3.9.6
pydantic==2.8.2
